requirements.txt
discord.py==2.7.1
aiosqlite==0.22.1
python-dotenv==1.2.2
aiohttp==3.13.5
cachetools==7.0.6
